requirements.txt
flask
python-chess
google-genai
markdown
python-dotenv
